get compression to work, ! changes syscall args !
the new args include an offset, so the kumem is not fully written and the
offset and capacity are now passed in words (uint64_t), not bytes.
the reason: userside wants to do compression and has to insert an
compression header (is it compressed, where does data start, how long is
data (in bytes, not words!)). but it wants to add this directly in front
of the actual raw data in case compression is not useful. so the kumem
may start with e.g. four bytes of header and the kernel btb export is
informed where it may write in the kumem via the offset.