add rough fix for output of incorrect jdb_btb
the bug is: I added cpu_id to all entry types, even BTE_INFO.
this shifts all other entry attributes back.
but info can't describe itself, so this update has to hardcoded in.
it is now solved for future entries, by removing the cpu_id from just
BTE_INFO, but I wanted to keep this specific exported file, so it got
the version == 2 (in the .traced, what a hassle) and this padding can
soon be deleted again.